Improve structure and test coverage
