+++ mocks/emergency-banner.js
@@ -1,0 +1,12 @@
+export default `<div class="nhsuk-global-alert" id="nhsuk-global-alert">
+<div class="nhsuk-width-container">
+<div class="nhsuk-grid-row">
+<div class="nhsuk-grid-column-full">
+<div class="nhsuk-global-alert__content">
+<p class="nhsuk-global-alert__message">If you have recently travelled from China or are looking for information on the coronavirus outbreak please go to our page on <a href="https://www.nhs.uk/conditions/wuhan-novel-coronavirus/">coronavirus</a> or the information on <a href="https://gov.uk/coronavirus">GOV.UK</a></p>
+<p id="last-updated" class="nhsuk-global-alert__updated">Last updated on: 1 January 2020 at 10am</p>
+</div>
+</div>
+</div>
+</div>
+</div>`;
