+++ mocks/feedback-banner.js
@@ -1,0 +1,19 @@
+export default `<div class="nhsuk-feedback-banner" id="nhsuk-feedback-banner" style="display: block;">
+<div class="nhsuk-width-container">
+<div class="nhsuk-grid-row">
+<div class="nhsuk-grid-column-full">
+<div class="nhsuk-feedback-banner__content">
+<h2 class="nhsuk-feedback-banner__heading">Tell us what you think about this page</h2>
+<p class="nhsuk-feedback-banner__message nhsuk-u-margin-bottom-2">
+<a href="https://www.nhs.uk" id="AgreeButton" target="_blank">Take our survey (opens new page)</a>
+</p>
+<p class="nhsuk-feedback-banner__message nhsuk-u-margin-bottom-4">
+<a href="#" id="DisagreeButton">I do not want to take this survey</a>
+</p>
+<p class="nhsuk-u-margin-bottom-0">We"ll use a cookie to save your choice.</p>
+<button class="nhsuk-feedback-banner__close" id="nhsuk-feedback-banner-close" type="button">Hide<span class="nhsuk-u-visually-hidden"> feedback invite</span></button>
+</div>
+</div>
+</div>
+</div>
+</div>`;
