+++ src/js/templates.js
@@ -1,0 +1,32 @@
+export const feedbackBannerStr = `<div class="nhsuk-feedback-banner" id="nhsuk-feedback-banner">
+<div class="nhsuk-width-container">
+<div class="nhsuk-grid-row">
+<div class="nhsuk-grid-column-full">
+<div class="nhsuk-feedback-banner__content">
+<h2 class="nhsuk-feedback-banner__heading">[Invitation1]</h2>
+<p class="nhsuk-feedback-banner__message nhsuk-u-margin-bottom-2">
+<a href="[BannerUrl]" id="AgreeButton" target="_blank">Take our survey (opens new page)</a>
+</p>
+<p class="nhsuk-feedback-banner__message nhsuk-u-margin-bottom-4">
+<a href="#" id="DisagreeButton">I do not want to take this survey</a>
+</p>
+<p class="nhsuk-u-margin-bottom-0">We"ll use a cookie to save your choice.</p>
+<button class="nhsuk-feedback-banner__close" id="nhsuk-feedback-banner-close" type="button">Hide<span class="nhsuk-u-visually-hidden"> feedback invite</span></button>
+</div>
+</div>
+</div>
+</div>
+</div>`;
+
+export const emergencyBannerStr = `<div class="nhsuk-global-alert" id="nhsuk-global-alert">
+<div class="nhsuk-width-container">
+<div class="nhsuk-grid-row">
+<div class="nhsuk-grid-column-full">
+<div class="nhsuk-global-alert__content">
+<p class="nhsuk-global-alert__message">[Content]</p>
+<p id="last-updated" class="nhsuk-global-alert__updated"></p>
+</div>
+</div>
+</div>
+</div>
+</div >`;
