+++ src/scss/_emergency-alert.scss
@@ -8,7 +8,6 @@
  */
 
  @mixin nhsuk-link-style-inverted {
-
 	color: $nhsuk-link-color;
   
 	&:focus {
@@ -29,42 +28,36 @@
 	  background-color: $color_shade_nhsuk-blue-50;
 	  box-shadow: 0 0 0 4px $color_shade_nhsuk-blue-35;
 	  color: $color_nhsuk-white;
-	}
-  
-  }
+	}  
+}
 
- .nhsuk-global-alert {
+.nhsuk-global-alert {
 	@include print-hide();
-  
 	background-color: $color_nhsuk-yellow;
 	padding-bottom: nhsuk-spacing(4);
 	padding-top: nhsuk-spacing(4);
-  
+
 	a {
-	  @include nhsuk-link-style-inverted; /* [1] */
+		@include nhsuk-link-style-inverted; /* [1] */
 	}
+}
   
-  }
-  
-  .nhsuk-global-alert__heading {
-	@include nhsuk-typography-responsive(24);
-  
+.nhsuk-global-alert__heading {
+	@include nhsuk-typography-responsive(24);  
 	margin: 0;
-  }
+}
   
-  .nhsuk-global-alert__content {
-	@include top-and-bottom;
-  
+.nhsuk-global-alert__content {
+	@include top-and-bottom;  
 	position: relative;
-  }
+}
   
-  .nhsuk-global-alert__message {
-	@include nhsuk-responsive-margin(2, 'top');
-  
+.nhsuk-global-alert__message {
+	@include nhsuk-responsive-margin(2, 'top');  
 	margin-bottom: 0;
-  }
+}
   
-  .nhsuk-global-alert__updated {
+.nhsuk-global-alert__updated {
 	@include nhsuk-responsive-margin(2, 'top');
 	@include nhsuk-typography-responsive(14);
-  }
+}
