+++ src/scss/_feedback-banner.scss
@@ -14,7 +14,6 @@
 
  .nhsuk-feedback-banner {
 	@include print-hide();
-  
 	background-color: $color_nhsuk-white;
 	bottom: 0;
 	box-shadow: 0 -4px 0 0 $nhsuk-box-shadow;
@@ -25,42 +24,37 @@
 	position: fixed;
 	width: 100%;
 	z-index: 20; /* [2] */
-  
+
 	&.js-inview { /* [6] */
-	  bottom: auto;
-	  position: relative;
+		bottom: auto;
+		position: relative;
 	}
+}
   
-  }
-  
-  .nhsuk-feedback-banner__heading {
+.nhsuk-feedback-banner__heading {
 	@include nhsuk-typography-responsive(24);
-  
 	margin: 0;
 	padding-right: nhsuk-spacing(6); /* [3] */
-  }
+}
   
-  .nhsuk-feedback-banner__content {
+.nhsuk-feedback-banner__content {
 	@include clearfix();
-  
 	position: relative;
-  }
+}
   
-  .nhsuk-feedback-banner__message {
+.nhsuk-feedback-banner__message {
 	margin-bottom: 0;
 	margin-top: nhsuk-spacing(2);
 	padding-bottom: nhsuk-spacing(4); /* [4] */
-  
+
 	@include mq($from: mobile) {
-	  padding-bottom: 0;
+		padding-bottom: 0;
 	}
-  
-  }
-  
-  .nhsuk-feedback-banner__close {
+}
+
+.nhsuk-feedback-banner__close {
 	@include nhsuk-typography-responsive(14);
 	@include nhsuk-link-style-default;
-  
 	background: none;
 	border: 0;
 	color: $color_nhsuk-black;
@@ -70,13 +64,12 @@
 	right: 0;
 	text-decoration: underline;
 	top: 0;
-  
+
 	@include mq($until: mobile) {
-	  bottom: 0;
-	  position: absolute;
-	  right: 0;
-	  top: auto;
+		bottom: 0;
+		position: absolute;
+		right: 0;
+		top: auto;
 	}
-  
-  }
+}
