+++ src/scss/banner.scss
@@ -1,5 +1,7 @@
 @import 'node_modules/nhsuk-frontend/packages/core/settings/all';
 @import 'node_modules/nhsuk-frontend/packages/core/tools/all';
 
-@import './emergency-alert';
-@import './feedback-banner';
+#nhsuk-banner {
+  @import './emergency-alert';
+  @import './feedback-banner';
+}
